% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratObjectMaker-package.R
\docType{package}
\name{SeuratObjectMaker-package}
\alias{SeuratObjectMaker}
\alias{SeuratObjectMaker-package}
\title{SeuratObjectMaker: A Webapp For Turning A Count Matrix Directory Into A Seurat Object}
\description{
Launch a Shiny webapp. Load a count matrix directory, interactively filter on metadata, view results, download Seurat object and results of differential gene expression testing.
}
\author{
\strong{Maintainer}: Rasmus Olofzon \email{rasmus.olofzon@gmail.com} (\href{https://orcid.org/0000-0001-8079-8718}{ORCID})

}
\keyword{internal}
