% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{SeuratObjectMaker}
\alias{SeuratObjectMaker}
\title{Initialse the SSOMaker in a browser}
\usage{
SeuratObjectMaker(
  appDir = getwd(),
  port = getOption("shiny.port"),
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"),
  workerId = "",
  quiet = FALSE,
  display.mode = c("auto", "normal", "showcase"),
  test.mode = getOption("shiny.testmode", FALSE)
)
}
\description{
Start the SSOmaker in a browser so you can import data and process it.
}
